Extract client-side filtering logic into reusable utility module
Created a new filterUtils module in the datagrid directory that provides:
- applyFilterModel(): Main function to filter data based on Sippy's filter model
- evaluateFilter(): Single filter evaluation logic
- useFilteredData(): React hook for convenience (optional)

This removes code duplication between RegressedTestsPanel and
TriagedRegressionTestList, and makes the filtering logic reusable by any
component that needs client-side filtering with Sippy's advanced filter
system.

Benefits:
- DRY: Eliminates ~140 lines of duplicated code
- Maintainability: Single source of truth for filter logic
- Consistency: Ensures all components filter data the same way
- Documentation: Includes comprehensive documentation with examples

The utility supports:
- Multiple filters with AND/OR operators
- NOT modifier for negating filters
- All Sippy filter operators (contains, equals, startsWith, endsWith, etc.)
- Numeric comparison operators (>, >=, <, <=)
- Performance optimization via React.useMemo
